refactor: split phase3.rs into submodules for quality gate
1554-line phase3.rs scored F (0.0/100 TDG Structural), blocking
all commits repo-wide. Split into signing/encryption/quantize/profile
submodules. Also adds top-level gate shadow-alias to ci.yml so PRs
can actually merge (org ruleset requires check named literally "gate",
reusable workflow only produces "ci / gate").

Refs paiml/infra#PMAT-157
