+++ .github/workflows/ci.yml
@@ -16,3 +16,18 @@ jobs:
       repo: ${{ github.event.repository.name }}
     secrets: inherit
 
+  # Top-level gate: satisfies org ruleset which requires check named "gate".
+  # The reusable workflow produces "ci / gate" but rulesets need exact match on "gate".
+  gate:
+    runs-on: ubuntu-latest
+    needs: [ci]
+    if: always()
+    steps:
+      - name: Check required jobs
+        run: |
+          if [ "${{ needs.ci.result }}" != "success" ]; then
+            echo "ci failed: ${{ needs.ci.result }}"
+            exit 1
+          fi
+          echo "All required jobs passed"
+
